requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
gunicorn>=22.0
pydantic>=2.7
pydantic-settings>=2.3
python-jose[cryptography]>=3.3
PyYAML>=6.0
typer>=0.12
rich>=13.7
pyfiglet>=1.0.2
prometheus-client>=0.21
prometheus-fastapi-instrumentator>=7.0
opentelemetry-api>=1.25
opentelemetry-sdk>=1.25
opentelemetry-instrumentation-fastapi>=0.46b0

sqlalchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
redis>=5.0

httpx>=0.27
